requirement.txt
Flask==2.3.3
WeasyPrint==60.1
python-dotenv